requirements.txt
Flask
openai
langchain
python-dotenv
